requirements.txt
streamlit
pandas
python-dotenv
groq
chromadb
semantic-router
sentence-transformers
huggingface_hub
sqlite3